refactor(slack): md utils and consistent relative paths
