+++ backend/src/slack/LiveTopicMessage.ts
@@ -1,20 +1,16 @@
 import { groupBy, uniq } from "lodash";
 import { Blocks, Elements, Md, Message } from "slack-block-builder";
 
-import { createSlackLink } from "~backend/src/notifications/sendNotification";
-import { slackClient } from "~backend/src/slack/app";
-import { generateMarkdownFromTipTapJson } from "~backend/src/slack/slackMarkdown/generator";
-import {
-  REQUEST_TYPE_EMOJIS,
-  fetchTeamBotToken,
-  fetchTeamMemberBotToken,
-  findSlackUserId,
-} from "~backend/src/slack/utils";
 import { Topic, db } from "~db";
 import { RichEditorNode } from "~richEditor/content/types";
 import { assert, assertDefined } from "~shared/assert";
 import { routes } from "~shared/routes";
 import { MENTION_TYPE_LABELS, RequestType } from "~shared/types/mention";
+
+import { slackClient } from "./app";
+import { generateMarkdownFromTipTapJson } from "./md/generator";
+import { createSlackLink } from "./md/utils";
+import { REQUEST_TYPE_EMOJIS, fetchTeamBotToken, fetchTeamMemberBotToken, findSlackUserId } from "./utils";
 
 export async function LiveTopicMessage(topic: Topic) {
   const [message, topicMembers] = await Promise.all([
