+++ backend/src/slack/md/generator.test.ts
@@ -1,5 +1,5 @@
-import { generateMarkdownFromTipTapJson } from "~backend/src/slack/slackMarkdown/generator";
-import { parseAndTransformToTipTapJSON } from "~backend/src/slack/slackMarkdown/parser";
+import { generateMarkdownFromTipTapJson } from "./generator";
+import { parseAndTransformToTipTapJSON } from "./parser";
 
 const advancedMessage = `hello <https://www.google.com/> :rolling_on_the_floor_laughing: test <#C02D6BU8J6P|general>
 >  I have a dream
